--- a/deeplabcutcore/pose_estimation_tensorflow/nnet/losses.py
+++ b/deeplabcutcore/pose_estimation_tensorflow/nnet/losses.py
@@ -4,7 +4,7 @@ https://github.com/eldar/pose-tensorflow
 '''
 import tensorflow as tf
 vers = (tf.__version__).split('.')
-if int(vers[0])==1 and int(vers[1])>12:
+if int(vers[0])==2 or int(vers[0])==1 and int(vers[1])>12:
     TF=tf.compat.v1
 else:
     TF=tf