--- a/deeplabcutcore/pose_estimation_tensorflow/nnet/predict.py
+++ b/deeplabcutcore/pose_estimation_tensorflow/nnet/predict.py
@@ -19,7 +19,7 @@ https://arxiv.org/abs/1909.11229
 import numpy as np
 import tensorflow as tf
 vers = (tf.__version__).split('.')
-if int(vers[0])==1 and int(vers[1])>12:
+if int(vers[0])==2 or int(vers[0])==1 and int(vers[1])>12:
     TF=tf.compat.v1
 else:
     TF=tf
@@ -183,16 +183,16 @@ def getposeNP(image, cfg, sess, inputs, outputs, outall=False):
 
 ### Code for TF inference on GPU
 def setup_GPUpose_prediction(cfg):
-    tf.compat.v1.reset_default_graph()
-    inputs = tf.compat.v1.placeholder(tf.float32, shape=[cfg.batch_size   , None, None, 3])
+    TF.reset_default_graph()
+    inputs = TF.placeholder(TF.float32, shape=[cfg.batch_size   , None, None, 3])
     net_heads = pose_net(cfg).inference(inputs)
     outputs = [net_heads['pose']]
 
-    restorer = tf.compat.v1.train.Saver()
-    sess = tf.compat.v1.Session()
+    restorer = TF.train.Saver()
+    sess = TF.Session()
 
-    sess.run(tf.compat.v1.global_variables_initializer())
-    sess.run(tf.compat.v1.local_variables_initializer())
+    sess.run(TF.global_variables_initializer())
+    sess.run(TF.local_variables_initializer())
 
     # Restore variables from disk.
     restorer.restore(sess, cfg.init_weights)