https://github.com/DeepLabCut/DeepLabCut-core/issues/6#issuecomment-694903268